финальная версия задания на первичную проверку
