+++ src/mock/comment.js
@@ -1,7 +1,7 @@
 import dayjs from 'dayjs';
 import {nanoid} from 'nanoid';
 import {DESCRIPTION_SENTENCES, PERSONS, ID_LENGTH, DateInMillisecondsCount} from './constant.js';
-import {getRandomInteger, getRandomValue} from '../util.js';
+import {getRandomInteger, getRandomValue} from '../util/common.js';
 
 
 const EMOTIONS = ['smile', 'sleeping', 'puke', 'angry'];
