+++ src/view/abstract.js
@@ -1,4 +1,4 @@
-import {createElement} from '../util.js';
+import {createElement} from '../util/render.js';
 
 
 export default class Abstract {
